Convert request paths to file paths
The server configuration for `web-component-tester` was changed so
that all files are served under the `/components` path.  A lookup
table is used to map from request path to filesystem path.  In
order to instrument files, we need the filesystem path, so this
replicates the functionality of the `serve-waterfall` module in
order to build a lookup table based on `web-component-tester`
configuration.